+++ cogs/_cselfintroduction.py
@@ -4,10 +4,10 @@ from discord.ext import commands
 # CSelfintroduction
 
 
-class CSelfintroduction(commands.Cog):
+class CSelfIntroduction(commands.Cog):
     def __init__(self, bot: commands.Bot):
         self.bot = bot
 
 
 async def setup(bot: commands.Bot):
-    await bot.add_cog(CSelfintroduction(bot))
+    await bot.add_cog(CSelfIntroduction(bot))
