+++ cogs/cbnoticetime.py
@@ -5,19 +5,6 @@ from datetime import datetime, timedelta
 
 from discord import app_commands
 from config.config import config
-from discord.abc import User
-
-
-class CustomBot(commands.Bot):
-    async def is_owner(self, user: User) -> bool:
-        return user.id in config.owner_ids
-
-
-client = CustomBot(
-    intents=discord.Intents.all(),
-    command_prefix=config.prefix,
-    owner_ids=config.owner_ids
-)
 
 
 class CBnoticetime(commands.Cog):
